Cleanning and place ugly graphical utils where it should lies
