+++ apps/activity/templates/activity/activity_type/pltp/small.html
@@ -1,6 +1,5 @@
 <ion-card>
   <!-- <ion-img src="{{ static('activity/images/pltp.png') }}" width="10%"></ion-img> -->
-  <!-- <img src="{{ static('activity/images/pltp.png') }}" width="10%"> -->
   <table>
     <tr>
       <td width="50%"><img src="{{ static('activity/images/pltp.png') }}" ></td>
@@ -15,7 +14,7 @@
     </tr>
     <tr>
       <td align="center"><b>Avancement<br>sur les exos</b></td>
-      <td align="center"><b>Moyenne des<br>exos traités</b></td>
+      <td align="center"><b>Réussite sur<br>exos traités</b></td>
     </tr>
     <tr>
       <td align="center">{{ progr|safe }}</td><td align="center">{{ quality|safe }}</td>
